% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBoilerplate.R
\name{createBoilerplate}
\alias{createBoilerplate}
\title{Title}
\usage{
createBoilerplate(name, path = ".", webapp = TRUE, webapp_name = name)
}
\arguments{
\item{webapp_name}{}
}
\description{
Title
}
